+++ LaClock/ModSettings/LocaleEN.cs
@@ -22,7 +22,10 @@ namespace LaClock
                 { m_Setting.GetOptionGroupLocaleID(ModSettings.kFormatGroup), "Formatting" },
                 { m_Setting.GetOptionGroupLocaleID(ModSettings.kFrictionGroup), "Friction" },
 
-                { m_Setting.GetOptionLabelLocaleID(nameof(ModSettings.ClockFormatString)), "Clock Formatting" },
+                { m_Setting.GetOptionLabelLocaleID(nameof(ModSettings.ClockFormatChoice)), "Clock Formatting" },
+                { m_Setting.GetOptionDescLocaleID(nameof(ModSettings.ClockFormatChoice)), "Pick a format for the clock." },
+
+                { m_Setting.GetOptionLabelLocaleID(nameof(ModSettings.ClockFormatString)), "Clock Formatting (Custom...)" },
                 { m_Setting.GetOptionDescLocaleID(nameof(ModSettings.ClockFormatString)), @$"Display Format of the Clock.
 
 Examples:
@@ -42,14 +45,21 @@ Note: You can also surround parts with double asterisks to make them **\*\*bold\
                 { m_Setting.GetOptionLabelLocaleID(nameof(ModSettings.EnableBlink)), "Enable blink" },
                 { m_Setting.GetOptionDescLocaleID(nameof(ModSettings.EnableBlink)), "Enable blinking when below conditions are met." },
 
-                { m_Setting.GetOptionLabelLocaleID(nameof(ModSettings.BlinkCondition)), "Blink when:" },
-                { m_Setting.GetOptionDescLocaleID(nameof(ModSettings.BlinkCondition)), "How often to make the clock blink?" },
+                { m_Setting.GetOptionLabelLocaleID(nameof(ModSettings.BlinkPerMin)), "Blink every..." },
+                { m_Setting.GetOptionDescLocaleID(nameof(ModSettings.BlinkPerMin)), "How often to make the clock blink?" },
 
-                { m_Setting.GetEnumValueLocaleID(ModSettings.FrictionTriggerEnum.Never), "Never" },
-                { m_Setting.GetEnumValueLocaleID(ModSettings.FrictionTriggerEnum.Per60min), "Every hour" },
-                { m_Setting.GetEnumValueLocaleID(ModSettings.FrictionTriggerEnum.Per30min), "Every 30 minutes" },
-                { m_Setting.GetEnumValueLocaleID(ModSettings.FrictionTriggerEnum.Per1min), "Every minute :-p" },
-                { m_Setting.GetEnumValueLocaleID(ModSettings.FrictionTriggerEnum.Always), "Always??" },
+                { m_Setting.GetOptionLabelLocaleID(nameof(ModSettings.BlinkDurationSec)), "Blink duration (in seconds)" },
+                { m_Setting.GetOptionDescLocaleID(nameof(ModSettings.BlinkDurationSec)), "How many seconds to blink?" },
+
+
+
+                { m_Setting.GetEnumValueLocaleID(ModSettings.ClockFormatEnum.Custom), "(Custom...)" },
+                { m_Setting.GetEnumValueLocaleID(ModSettings.ClockFormatEnum.St), ModSettings.kExampleDateTime.ToString("t") },
+                { m_Setting.GetEnumValueLocaleID(ModSettings.ClockFormatEnum.Sts), ModSettings.kExampleDateTime.ToString("T") },
+                { m_Setting.GetEnumValueLocaleID(ModSettings.ClockFormatEnum.Sg), ModSettings.kExampleDateTime.ToString("g") },
+                { m_Setting.GetEnumValueLocaleID(ModSettings.ClockFormatEnum.Sgs), ModSettings.kExampleDateTime.ToString("G") },
+                { m_Setting.GetEnumValueLocaleID(ModSettings.ClockFormatEnum.HoM1), ModSettings.kExampleDateTime.ToString("**HH:mm** | ddd dd MMM") },
+                { m_Setting.GetEnumValueLocaleID(ModSettings.ClockFormatEnum.HoM1s), ModSettings.kExampleDateTime.ToString("**HH:mm:ss** | ddd dd MMM") },
 
             };
         }